Monitor workflow files in non-default locations. (#69148)
In addition to the default `.github/workflows` directory for GitHub Actions workflows, the repo also has a `setup-node` folder containing a composite action. This is not currently monitored by Dependabot for updates.

There are several updates to 3rd party actions in this composite action that are needed to avoid problems as a result of upstream changes. This updates the Dependabot configuration to also monitor this directory.

Co-authored-by: desrosj <desrosj@git.wordpress.org>